Add Run History to /candidate-test-plan/
