+++ client/components/Reports/SummarizeTestPlanReport.jsx
@@ -19,6 +19,7 @@ import TestPlanResultsTable from '../common/TestPlanResultsTable';
 import DisclosureComponent from '../common/DisclosureComponent';
 import { Link, Navigate, useLocation, useParams } from 'react-router-dom';
 import createIssueLink from '../../utils/createIssueLink';
+import { getTestersRunHistory } from './getTestersRunHistory';
 
 const ResultsContainer = styled.div`
   padding: 1em 1.75em;
@@ -27,54 +28,6 @@ const ResultsContainer = styled.div`
   border-bottom: 1px solid #dee2e6;
   margin-bottom: 0.5em;
 `;
-
-const getTestersRunHistory = (
-  testPlanReport,
-  testId,
-  draftTestPlanRuns = []
-) => {
-  const { id: testPlanReportId, at, browser } = testPlanReport;
-  let lines = [];
-
-  draftTestPlanRuns.forEach(draftTestPlanRun => {
-    const { testPlanReport, testResults, tester } = draftTestPlanRun;
-
-    const testResult = testResults.find(item => item.test.id === testId);
-    if (testPlanReportId === testPlanReport.id && testResult?.completedAt) {
-      lines.push(
-        <li
-          key={`${testResult.atVersion.id}-${testResult.browserVersion.id}-${testResult.test.id}-${tester.username}`}
-        >
-          Tested with{' '}
-          <b>
-            {at.name} {testResult.atVersion.name}
-          </b>{' '}
-          and{' '}
-          <b>
-            {browser.name} {testResult.browserVersion.name}
-          </b>{' '}
-          by{' '}
-          <b>
-            <a href={`https://github.com/${tester.username}`}>
-              {tester.username}
-            </a>
-          </b>{' '}
-          on {convertDateToString(testResult.completedAt, 'MMMM DD, YYYY')}.
-        </li>
-      );
-    }
-  });
-
-  return (
-    <ul
-      style={{
-        marginBottom: '0'
-      }}
-    >
-      {lines}
-    </ul>
-  );
-};
 
 const SummarizeTestPlanReport = ({ testPlanVersion, testPlanReports }) => {
   const { exampleUrl, designPatternUrl } = testPlanVersion.metadata;
